Refactor: 이전 ladder, line 모두 삭제하고 새로운 Ladder, LadderLine으로 동작하도록 함
