+++ src/main/java/nextstep/ladder/LadderApplication.java
@@ -2,8 +2,8 @@ package nextstep.ladder;
 
 import nextstep.ladder.domain.EndPoints;
 import nextstep.ladder.domain.GameResult;
-import nextstep.ladder.domain.Ladder;
 import nextstep.ladder.domain.Length;
+import nextstep.ladder.domain.Ladder;
 import nextstep.ladder.domain.Players;
 import nextstep.ladder.view.InputView;
 import nextstep.ladder.view.ResultView;
@@ -13,7 +13,7 @@ public class LadderApplication {
   public static void main(String[] args) {
     Players players = Players.of(InputView.getNames());
     EndPoints endPoints = EndPoints.of(InputView.getResults(), players.size());
-    Ladder ladder = Ladder.of(new Length(players.size()), new Length(InputView.getHeight()));
+    Ladder ladder = Ladder.init(new Length(players.size()), new Length(InputView.getHeight()));
     ResultView.printLadder(players, ladder, endPoints);
 
     GameResult gameResult = play(players, endPoints, ladder);
