--- src/main/java/nextstep/ladder/domain/BooleanGeneratingStrategy.java
@@ -1,6 +1,0 @@
-package nextstep.ladder.domain;
-
-@FunctionalInterface
-public interface BooleanGeneratingStrategy {
-  boolean getBoolean();
-}
